backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
ultralytics==8.1.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2